product

+++ app/src/main/java/com/hema/e_commerce/model/dataclass/allProducts/Option.kt
@@ -1,9 +1,9 @@
 package com.hema.e_commerce.model.dataclass.allProducts
 
 data class Option(
-    val id: Int,
+    val id: Double,
     val name: String,
     val position: Int,
-    val product_id: Int,
+    val product_id: Double,
     val values: List<String>
 )
